fix: support I64, U32, U64 writing and reading

+++ src/main/java/org/molgenis/vipannotate/annotation/IntAnnotationDecoder.java
@@ -10,13 +10,13 @@ public class IntAnnotationDecoder implements AnnotationDecoder<IntAnnotation> {
 
   @Override
   public IntAnnotation decode(MemoryBuffer memBuffer, int annotationIndex) {
-    int value = intReadValueFunction.apply(memBuffer, annotationIndex);
+    long value = intReadValueFunction.apply(memBuffer, annotationIndex);
     return new IntAnnotation(value);
   }
 
   @Override
   public void decodeInto(MemoryBuffer memBuffer, int annotationIndex, IntAnnotation annotation) {
-    int value = intReadValueFunction.apply(memBuffer, annotationIndex);
+    long value = intReadValueFunction.apply(memBuffer, annotationIndex);
     annotation.setValue(value);
   }
 }
